+++ packages/collaboration/comments/src/types.ts
@@ -1,8 +1,11 @@
+import { Node } from 'slate';
 import { Thread } from './Thread';
 
 export interface ThreadNodeData {
   thread: Thread;
   selected: boolean;
 }
+
+export type ThreadNode = Node & ThreadNodeData;
 
 export interface ThreadPlugin {}
