+++ packages/ui/collaboration/comments/src/AddThreadToolbarButton.tsx
@@ -1,8 +1,5 @@
 import React from 'react';
-import { ELEMENT_THREAD } from '@udecode/plate-comments';
 import {
-  getPluginType,
-  someNode,
   useEventPlateId,
   usePlateEditorState,
   withPlateEventProvider,
@@ -14,15 +11,12 @@ export const AddThreadToolbarButton = withPlateEventProvider(
     id = useEventPlateId(id);
     const editor = usePlateEditorState(id)!;
 
-    const type = getPluginType(editor, ELEMENT_THREAD);
-    const isThread =
-      !!editor?.selection && someNode(editor, { match: { type } });
-
     return (
       <ToolbarButton
-        active={isThread}
         onMouseDown={(event) => {
-          if (!editor) return;
+          if (!editor) {
+            return;
+          }
 
           event.preventDefault();
           onAddThread();
