--- packages/ui/collaboration/comments/src/AddThreadToolbarButton/index.ts
@@ -1,1 +1,0 @@
-export * from './AddThreadToolbarButton';
