--- packages/ui/collaboration/comments/src/SideThread/SideThreadComment/index.ts
@@ -1,1 +1,0 @@
-export * from './SideThreadComment';
