+++ packages/ui/collaboration/comments/src/Thread/ThreadComment/index.ts
@@ -1,0 +1,1 @@
+export * from './ThreadComment';
