+++ packages/ui/collaboration/comments/src/Thread/index.ts
@@ -1,0 +1,2 @@
+export * from './Thread';
+export * from './Thread.styles';
