+++ packages/ui/collaboration/comments/src/Threads/index.ts
@@ -1,0 +1,2 @@
+export * from './Threads';
+export * from './Threads.styles';
