+++ packages/ui/collaboration/comments/src/determineAbsolutePosition.tsx
@@ -1,0 +1,14 @@
+export function determineAbsolutePosition(element: HTMLElement) {
+  let left = 0;
+  let top = 0;
+  let currentElement: HTMLElement | null = element;
+  do {
+    left += currentElement.offsetLeft || 0;
+    top += currentElement.offsetTop || 0;
+    currentElement = currentElement.offsetParent as HTMLElement;
+  } while (currentElement);
+  return {
+    left,
+    top,
+  };
+}
