+++ packages/ui/collaboration/comments/src/index.ts
@@ -1,4 +1,6 @@
-export * from './AddThreadToolbarButton/index';
+export * from './AddThreadToolbarButton';
+export * from './ToggleShowThreadsButton';
 export * from './ThreadElement/index';
 export * from './SideThread/index';
 export * from './useComments';
+export { determineAbsolutePosition } from './determineAbsolutePosition';
